done with the section and subsections crud on ui took 3 days for this and controllers were also changed due to not giving course in response which caused course updation and fetching on ui impossible

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection=require("../models/SubSection")
 
 //createsection
 exports.createSection=async(req , res)=>{
@@ -25,7 +26,12 @@ exports.createSection=async(req , res)=>{
       $push:{
         courseContent:newSection._id
       },
-    },{new:true})
+    },{new:true}).populate({
+      path:"courseContent",
+      populate:{
+        path:"subSection"
+      }
+    })
 
     return res.status(200).json({
       success:true,
@@ -51,7 +57,7 @@ exports.updateSection=async(req ,res)=>{
     //update data
     //course has the section id so we dont need to update in course
 
-    const{sectionName,sectionId}=req.body;
+    const{sectionName,sectionId,courseId}=req.body;
     if(!sectionName || !sectionId){
       return res.status(400).json({
         success:false,
@@ -63,9 +69,19 @@ exports.updateSection=async(req ,res)=>{
       {sectionName},{new:true}
     )
 
+    const course=await Course.findById(courseId)
+      .populate({
+        path:"courseContent",
+        populate:{
+          path:"subSection"
+        }
+      })
+      .exec()
+
     return res.status(200).json({
       success:true,
-      message:"Section updated successfully"
+      message:"Section updated successfully",
+      data:course
     })
 
   }
@@ -78,6 +94,7 @@ exports.updateSection=async(req ,res)=>{
   }
 }
 
+
 //deleteSection
 exports.deleteSection=async(req ,res)=>{
   try{
@@ -86,21 +103,51 @@ exports.deleteSection=async(req ,res)=>{
     //course object id remove=>not needed
     //return response
 
-    const {sectionId}=req.params;
+    const {sectionId}=req.body;
+    const {courseId}=req.body;
+
+
+    await Course.findByIdAndUpdate(courseId,{
+      $pull:{
+        courseContent:sectionId,
+      },
+    })
+
+    const section=await Section.findById(sectionId)
+
+    if(!section){
+      return res.status(404).json({
+        success:false,
+        message:"Section not found"
+      })
+    }
+
+    await SubSection.deleteMany({_id:{$in:section.subSection || []}})
+
     await Section.findByIdAndDelete(sectionId);
 
+    const course=await Course.findById(courseId)
+      .populate({
+        path:"courseContent",
+        populate:{
+          path:"subSection"
+        }
+      })
+      .exec()
 
     return res.status(200).json({
       success:true,
-      message:"Section deleted"
+      message:"Section deleted",
+      data:course
     })
 
   }
   catch(error){
-    return res.status(500).json({
-      success:false,
-      message:"Unable to delete Section, please try again later",
-      error:error.message
-    })
+      console.log("DELETE SECTION ERROR:",error);
+      return res.status(500).json({
+        success:false,
+        message:"Unable to delete Section, please try again later",
+        error:error.message
+      })
   }
 }
